--- advent_of_code/aoc/2022/inputs/day23_sample.txt
@@ -1,7 +1,0 @@
-....#..
-..###.#
-#...#.#
-.#...##
-#.###..
-##.#.##
-.#..#..
